+++ lib/subatomic/icons.js
@@ -133,7 +133,7 @@ import refresh from './icons/refresh';
 import refreshSolid from './icons/refreshSolid';
 import energy from './icons/energy';
 import energySolid from './icons/energySolid';
-import aiStars from './icons/aiStars';
+import aiStarsSolid from './icons/aiStarsSolid';
 
 export default {
 	base: function (icon) {
@@ -695,5 +695,5 @@ export default {
 	refreshSolid,
 	energy,
 	energySolid,
-	aiStars
+	aiStarsSolid
 };
