+++ lib/subatomic/icons/aiStarsSolid.js
